Simplify string normalization using modern JS functions
Simplify the utility for normalizing strings now that
`String.prototype.normalize` and `\p` escapes are widely available.
