+++ h/static/scripts/tests/util/string-test.js
@@ -1,26 +1,26 @@
-import * as stringUtil from '../../util/string';
+import { hyphenate, unhyphenate, stripMarks } from '../../util/string';
 
 describe('util/string', () => {
   describe('hyphenate', () => {
     it('converts input to kebab-case', () => {
-      assert.equal(stringUtil.hyphenate('fooBar'), 'foo-bar');
-      assert.equal(stringUtil.hyphenate('FooBar'), '-foo-bar');
+      assert.equal(hyphenate('fooBar'), 'foo-bar');
+      assert.equal(hyphenate('FooBar'), '-foo-bar');
     });
   });
 
   describe('unhyphenate', () => {
     it('converts input to camelCase', () => {
-      assert.equal(stringUtil.unhyphenate('foo-bar'), 'fooBar');
-      assert.equal(stringUtil.unhyphenate('foo-bar-'), 'fooBar');
-      assert.equal(stringUtil.unhyphenate('foo-bar-baz'), 'fooBarBaz');
-      assert.equal(stringUtil.unhyphenate('-foo-bar-baz'), 'FooBarBaz');
+      assert.equal(unhyphenate('foo-bar'), 'fooBar');
+      assert.equal(unhyphenate('foo-bar-'), 'fooBar');
+      assert.equal(unhyphenate('foo-bar-baz'), 'fooBarBaz');
+      assert.equal(unhyphenate('-foo-bar-baz'), 'FooBarBaz');
     });
   });
 
-  describe('stringUtil helpers', () => {
+  describe('stripAccents', () => {
     it('removes hungarian marks', () => {
       const text = 'Fürge rőt róka túlszökik zsíros étkű kutyán';
-      const decoded = stringUtil.fold(stringUtil.normalize(text));
+      const decoded = stripMarks(text);
       const expected = 'Furge rot roka tulszokik zsiros etku kutyan';
 
       assert.equal(decoded, expected);
@@ -28,7 +28,7 @@ describe('util/string', () => {
 
     it('removes greek marks', () => {
       const text = 'Καλημέρα κόσμε';
-      const decoded = stringUtil.fold(stringUtil.normalize(text));
+      const decoded = stripMarks(text);
       const expected = 'Καλημερα κοσμε';
 
       assert.equal(decoded, expected);
@@ -36,7 +36,7 @@ describe('util/string', () => {
 
     it('removes japanese marks', () => {
       const text = 'カタカナコンバータ';
-      const decoded = stringUtil.fold(stringUtil.normalize(text));
+      const decoded = stripMarks(text);
       const expected = 'カタカナコンハータ';
 
       assert.equal(decoded, expected);
@@ -44,7 +44,7 @@ describe('util/string', () => {
 
     it('removes marathi marks', () => {
       const text = 'काचं शक्नोम्यत्तुम';
-      const decoded = stringUtil.fold(stringUtil.normalize(text));
+      const decoded = stripMarks(text);
       const expected = 'कच शकनमयततम';
 
       assert.equal(decoded, expected);
@@ -52,7 +52,7 @@ describe('util/string', () => {
 
     it('removes thai marks', () => {
       const text = 'ฉันกินกระจกได้ แต่มันไม่ทำให้ฉันเจ็บ';
-      const decoded = stringUtil.fold(stringUtil.normalize(text));
+      const decoded = stripMarks(text);
       const expected = 'ฉนกนกระจกได แตมนไมทาใหฉนเจบ';
 
       assert.equal(decoded, expected);
@@ -61,7 +61,7 @@ describe('util/string', () => {
     it('removes all marks', () => {
       const text =
         '̀ ́ ̂ ̃ ̄ ̅ ̆ ̇ ̈ ̉ ̊ ̋ ̌ ̍ ̎ ̏ ̐ ̑ ̒ ̓ ̔ ̕ ̖ ̗ ̘ ̙ ̚ ̛ ̜ ̝ ̞ ̟ ̠ ̡ ̢ ̣ ̤ ̥ ̦ ̧ ̨ ̩ ̪ ̫ ̬ ̭ ̮ ̯ ̰ ̱ ̲ ̳ ̴ ̵ ̶ ̷ ̸ ̹ ̺ ̻ ̼ ̽ ̾ ̿ ̀ ́ ͂ ̓ ̈́ ͅ ͠ ͡"';
-      const decoded = stringUtil.fold(stringUtil.normalize(text));
+      const decoded = stripMarks(text);
       const expected =
         '                                                                       "';
